requirements.txt
functions-framework==3.9.1
google-genai>=1.12.1
google-api-python-client>=2.164.0
google-auth>=2.38.0
google-cloud-firestore>=2.16.0
pypdf>=5.3.0
python-docx>=1.1.2